+++ CHANGELOG.md
@@ -4,7 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- Added an Equality comparison overload to ByteArray
+
+### Changed
+- Updated Readme.md with examples
+
+
 ## [0.1.0-alpha.1]
+
 ### Added
 - Added a new copy constructor that copies only n bytes as given in the constructor itself